Added a no_bar option to OOFMenuItem.
This prevents a submenu from appearing in a window's menu bar.
Most menus that had no_gui==True now have no_bar==True instead.

Added the ability to add OOFMenuItem options that are not
recursive.  Setting it in one menu doesn't affect the
menu's submenus or menu items.  no_bar is currently the
only nonrecursive menu option.

Added a verbose flag to some menu operations for debugging.
It will be removed in a later commit.